add node-fetch and dynamic list rendering with react app
