+++ app.js
@@ -17,8 +17,7 @@ const dataRouter = require('./routes/data')
 
 // express configurations
 app.set('view engine', 'pug')
-app.use(express.static(path.join(__dirname, "public")))
-// app.use(express.static("public"))
+app.use(express.static(path.join(__dirname, 'public')))
 app.use(morgan('dev'))
 app.use(cookieParser(sessionSecret))
 app.use(express.urlencoded({ extended: false }))
