+++ package.json
@@ -28,19 +28,19 @@
     "express-session": "^1.17.1",
     "faker": "^4.1.0",
     "morgan": "^1.10.0",
+    "node-fetch": "^2.6.0",
     "path": "^0.12.7",
     "per-env": "^1.0.2",
     "pg": "^8.2.1",
     "pug": "^2.0.4",
-    "sequelize": "^5.21.12",
     "react": "^16.13.1",
     "react-beautiful-dnd": "^13.0.0",
     "react-dom": "^16.13.1",
     "react-scripts": "3.4.1",
+    "sequelize": "^5.21.12",
     "styled-components": "^5.1.1"
   },
   "devDependencies": {
-    "babel-plugin-transform-react-pug": "^7.0.1",
     "dotenv-cli": "^3.1.0",
     "nodemon": "^2.0.4",
     "sequelize-cli": "^5.5.1"
