--- public/indexs.html
@@ -1,9 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <title>C.A.R.L.</title>
-  </head>
-  <body>
-    <div id="root"></div>
-  </body>
-</html>
