+++ views/columns/columns.pug
@@ -7,8 +7,8 @@ block content
   script(src="https://unpkg.com/react-dom@16.8.6/umd/react-dom.development.js" crossorigin)
   script(src="https://unpkg.com/styled-components@4.1.1/dist/styled-components.js")
   script(src="https://unpkg.com/react-beautiful-dnd@13.0.0/dist/react-beautiful-dnd.js")
-  script(src='../../../../src/index.js' type="text/babel")
-
+  script(src='../../../../js/index.js' type="text/babel" state=state)
+  
 
 
   //- div.wrapper
